many imporvements

+++ src/state/linky.rs
@@ -4,17 +4,17 @@ use embassy_sync::mutex::Mutex;
 #[derive(Clone, Copy)]
 pub struct Linky {
     pub east: u32,
-    pub sinsts: u32,
+    pub sinsts: u16,
 }
 
-static LINKY: Mutex<CriticalSectionRawMutex, Linky> = Mutex::new(Linky{east: 0u32, sinsts: 0u32});
+static LINKY: Mutex<CriticalSectionRawMutex, Linky> = Mutex::new(Linky{east: 0u32, sinsts: 0u16});
 
 pub async fn set_east(east: u32) {
     let mut l = LINKY.lock().await;
     l.east = east;
 }
 
-pub async fn set_sinsts(sinsts: u32) {
+pub async fn set_sinsts(sinsts: u16) {
     let mut l = LINKY.lock().await;
     l.sinsts = sinsts;
 }
